research-development/aware_os/pyproject.toml
[project]
name = "alignncg"
version = "0.1.0"
description = "Alignment-based spectral triple sandbox (v3/v5 backends) with CLI"
requires-python = ">=3.10"
dependencies = [
  "typer[all]>=0.12",
  "numpy>=1.26",
  "scipy>=1.11",
  "pyyaml>=6.0",
  "rich>=13.7",
]

[project.scripts]
alignncg = "alignncg.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["./aware_os"]
